requirements.txt
opencv-python==4.5.5.64
face-recognition==1.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
dlib<20.1
flask==3.1.2
insightface
onnxruntime
pillow
easyocr
